I had corrected the way asserted should be done. Created a new TextBuddyTest to do unit testing in TDD form. Due to that, i Had to created new function to return value of output to compare using assert "getData" and also changing the "search()" to return a String value to compare outpur using assert. Additionally, main() had been splitted into three parts so as to allow assert unit testing -->>>
	public TextBuddy(String[] args){
		sc = new Scanner(System.in);        
		String fileName = args[0];
		file = new File(fileName);
		System.out.println(MESSAGE_WELCOME + file.getName() +
MESSAGE_READY_TO_USE);           	
	}

	public static void main(String[] args) {
		TextBuddy tb = new TextBuddy(args);
		tb.startUp();
	}

	void startUp() {
		String userInput;
		for (;;) {//this loop will carry on till exit command is given
			System.out.print(MESSAGE_COMMAND_INPUT);
			userInput = sc.nextLine();
			executeUserInput(userInput);
		}
	}
For these part, one will be the constructor, the other will be startup.
So as to allow main() to create an object for unit testing. 
The startUp() is just to start off the program using the object, making
it OOP style.
Last of all, I had remove the function checkIfFileExist as it is
redundant due to user using the same old file and it cannot access.